+++ doc/whats_new.rst
@@ -107,8 +107,10 @@ Changelog
     Gaussian and sparse random projection matrix
     by `Olivier Grisel`_ and `Arnaud Joly`_.
 
-   - The :fun:`metrics.mean_squared_error` and :fun:`metrics.r2_score` metrics
-     support multioutput by `Arnaud Joly`_.
+   - Add the :fun:`metrics.mean_absolute_error` function which computes the
+     mean absolute error. The :fun:`metrics.mean_squared_error`,
+     :fun:`metrics.mean_absolute_error` and
+     :fun:`metrics.r2_score` metrics support multioutput by `Arnaud Joly`_.
 
 API changes summary
 -------------------